Added PerspectiveShooterCharacter and Controller
Set it as the default Pawn and Controller. It inherits from the normal ShooterCharacter so hopefully there won't be any issues. First perspective works decently, and I added a keybind (currently P) to swap between first and third person. The muzzle flash is a bit wonky when you move forward as it stays in the air, maybe could try parenting it to the gun?